Remove unused code
Remove some ancient level indicator support code, as well as unused
logic setting the cursor. It appears at some point a separate cursor
was used for the selection area, but this is no longer the case. The
current behavior of displaying a cross everywhere matches Excel, so
I don't feel bad about removing the unused logic.