Column resize fix
Currently if the mouse is moved far to the left of the minimum column
width, moving the mouse rightward will enlarge the column header, even
though the cursor is still to the left of the drag handle. Restoring
this line of code produces the expected behavior of enlarging the column
header only when the cursor has the same X coordinate as the drag
handle.