Use a single backing layer for the whole view
It's not really necessary for the subviews to have backing layers,
as they'll be composited into the layer of a shared ancestor. The
asynchronous drawing stuff is probably not needed as the table grid is
now quite performant. (In fact maybe layer backing is not necessary
at all at this point.)