Basic Find / Replace support
Implement a Find / Replace bar via the NSTextFinder mechanisms. I found
AppKit to be buggy when searching very large strings (or arrays of
strings), so use a "virtual" string class where each cell represents
one "character".

Matching is performed by retrieving the text value from
tableGrid:objectValueForColumn:row:. The "character" is matched if the
search string is present in the cell's string representation. Replacing
is done using tableGrid:setObjectValue:forColumn:row:. Note that
Find/Replace therefore only works on entire cells. That is, you can't
replace a substring within a cell.

NSTextFinder kicks off background threads for searching. We prematurely
terminate these threads by periodically checking to see whether the Find
bar is still visible. If it's not visible, either because the user
pressed Done or the table grid was hidden, the search is stopped.

A new findBarVisible property has been added MBTableGrid, which can call
up the find bar, or hide it away. It simply passes the messages onto
the scroll view (which is the Find bar's container).

Subclassing NSScrollView was necessary in order to prevent certain
Find options from appearing. When the "Starts With" or "Whole Word"
options are checked in the Find bar's menu, NSTextFinder fires up
the internal regular expression engine, which is incompatible with
the virtual string strategy described above. So we hide those options
before displaying the Find bar by manipulating the searchMenuTemplate
of the associated NSSearchField.

Searching happens in column-first order, which is what I needed for
my own application. Row-first searching could be added later. Inquire
within.

For large grids, we'll likely need some kind of bulk-replacement
mechanism, e.g. a new setObjectValue: delegate method that takes an
NSIndexSet (or two). This should be considered in conjunction with a
revised deleteBackward: action, which currently loops over all of the
selected cells, and is very slow.