New data source method for bulk replacements / deletions
New optional callback method:

- (void)tableGrid:(MBTableGrid *)aTableGrid setObjectValue:(id)object
       forColumns:(NSIndexSet *)columns rows:(NSIndexSet *)rows;

This can be provided in addition to the existing singular form, or
in place of it. The purpose is to improve the efficiency of bulk
operations, such as pressing Delete on a selected block of cells.
When #38 is merged, it can also be used to implement "Replace All".

Fixes #23