Performance improvement
updateTrackingAreas was O(N^2) because after clearing the rect cache,
calculating each rect required fetching the width of every leftward
column. Now use recursion to fetch the cached location of the rect
immediate to the left, and use that to calculate the current rect.
This change makes updateTrackingAreas O(N).