New delegate methods
Delegate method that receives double-click events on individual cells:

    - (void)tableGrid:(MBTableGrid *)aTableGrid didDoubleClickColumn:(NSUInteger)columnIndex row:(NSUInteger)rowIndex;

Delegate method called before displaying a menu on the row footer:

    - (void)tableGrid:(MBTableGrid *)aTableGrid willDisplayFooterMenu:(NSMenu *)menu forRow:(NSUInteger)rowIndex;

These are modeled closely off of existing delegate methods.

This commit also consolidates rect-to-range logic in various places
with a new internal API on MBTableGrid:

    - (NSRange)_rangeOfColumnsIntersectingRect:(NSRect)rect;
    - (NSRange)_rangeOfRowsIntersectingRect:(NSRect)rect;

So future improvements in the column-finding algorithm will benefit
multiple parts of the code base.