Modernize grid and shadow colors
Use NSColor.shadowColor for shadows, and NSColor.quaternaryLabelColor
for the grid lines. NSColor.gridColor has changed drastically in Big
Sur, and is no longer usable in Dark Mode (it appears black).

Note that the quaternaryLabelColor docs state that can be used "for
separators between text items", so grid lines are a good candidate. The
color seems to be identical to NSColor.separatorColor, which was
introduced in Mojave.